% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_base.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calculer la distribution de l'âge}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{Un data frame contenant la date de naissance des élus.}
}
\value{
Un objet contenant les quantiles 0, 25, 50, 75, 100 de l’age des élus.
}
\description{
Cette fonction  calcule les quantiles de l’age des élus.
}
\examples{
df <- data.frame(Date.de.naissance = c("01/01/1950", "15/07/1975", "10/03/1990"))
}
