% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_base.R
\name{compter_nombre_d_elus}
\alias{compter_nombre_d_elus}
\title{Compter le nombre d'élus}
\usage{
compter_nombre_d_elus(df)
}
\arguments{
\item{df}{Un data frame contenant une colonne "Libellé.de.la.fonction" indiquant le rôle de chaque élu.}
}
\value{
Un  nombre unique d’élus basés sur l’unicité du triplet nom/prénom/date de naissance.
}
\description{
Cette fonction compte le nombre d'élus dans une liste.
}
\examples{
df <- data.frame(Nom.de.l.élu = c("Dupont", "Martin", "Dupont"),
                 Prénom.de.l.élu = c("Jean", "Sophie", "Jean"),
                 Date.de.naissance = c("01/01/1980", "15/07/1975", "01/01/1980"))
}
