% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_classe.R
\name{creer_commune}
\alias{creer_commune}
\title{Création d'un objet de classe commune}
\usage{
creer_commune(df)
}
\arguments{
\item{df}{Un `data.frame` contenant les informations d'une commune.}
}
\value{
Un objet de classe `commune`.
}
\description{
Cette fonction transforme un `data.frame` en objet de classe `commune`, en s'assurant que les colonnes nécessaires sont présentes.
}
\details{
- Le `data.frame` doit contenir au moins les colonnes `Libellé.de.la.commune` et `Libellé.du.département`.
- La classe de l'objet en sortie est définie comme `commune`, ce qui permet l'utilisation de méthodes spécifiques comme `plot.commune`.
}
\examples{
\dontrun{
df <- data.frame(
  Libellé.de.la.commune = "Paris",
  Libellé.du.département = "Paris",
  Code.de.la.catégorie.socio.professionnelle = c("A", "B", "C"),
  stringsAsFactors = FALSE
)
commune_obj <- creer_commune(df)
class(commune_obj) # "commune"
}

}
