% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_classe.R
\name{creer_departement}
\alias{creer_departement}
\title{Création d'un objet de classe département}
\usage{
creer_departement(df)
}
\arguments{
\item{df}{: Un `data.frame` contenant les informations des communes d'un département.}
}
\value{
Un objet de classe `departement`.
}
\description{
Cette fonction transforme un `data.frame` en objet de classe `departement`, en s'assurant que les colonnes nécessaires sont présentes.
}
\details{
* Le `data.frame` doit contenir au moins les colonnes `Libellé.du.département` et `Libellé.de.la.commune`.
* La classe de l'objet en sortie est définie comme `departement`, ce qui permet l'utilisation de méthodes spécifiques comme `plot.departement`.
}
\examples{
\dontrun{
df_departement <- data.frame(
  Libellé.de.la.commune = c("Nantes", "Saint-Herblain"),
  Libellé.du.département = "Loire-Atlantique",
  Code.de.la.catégorie.socio.professionnelle = c("A", "B", "C"),
  stringsAsFactors = FALSE
)
departement_obj <- creer_departement(df_departement)
class(departement_obj)
}

}
\seealso{
\code{\link{creer_commune}}
}
