% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_generiques.R
\name{plot.departement}
\alias{plot.departement}
\title{Visualisation des professions des élus d'un département}
\usage{
\method{plot}{departement}(x, ...)
}
\arguments{
\item{x}{Un objet de type `departement`, qui est un `data.frame` contenant les informations des communes d'un département.}

\item{...}{Arguments supplémentaires (actuellement non utilisés).}
}
\value{
Un graphique `ggplot` affichant la distribution des 10 catégories socio-professionnelles les plus représentées.
}
\description{
Cette fonction génère un graphique en barres représentant la répartition des élus d'un département selon leur catégorie socio-professionnelle.
}
\details{
- Le dataframe fourni doit contenir un département avec plusieurs communes.
- Le graphique affiche les 10 codes professionnels les plus représentés en nombre d'élus.
- Le titre du graphique correspond au nom du département suivi du nombre de communes.
- L'axe des abscisses précise le département concerné.
}
\examples{
\dontrun{
plot.departement(df_mon_departement)
}

}
