% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_base.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Compter le nombre d'élus ayant le même code professionnel.}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{: Un data frame contenant les codes des catégories socioprofessionnelles des élus.}
}
\value{
Un bar chart contenant le nombre d'élus pour chaque code professionnel.
}
\description{
Cette fonction compte le nombre d'élus pour chaque code professionnel et le représentera avec un bar chart.
}
\examples{
df <- data.frame(Code.de.la.catégorie.socio.professionnelle = c("A", "B", "A", "C", "B", "A"))
}
\seealso{
\code{\link{compter_nombre_d_adjoints}}, \code{\link{compter_nombre_d_elus}}, \code{\link{calcul_distribution_age}}, \code{\link{trouver_l_elu_le_plus_age}}
}
