% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_base.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Trouver l'élu le plus âgé}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{Un data frame contenant le nom, le prénom, et la date de naissance des élus.}
}
\value{
Un objet contenant le nom, le prénom, et l’âge de l’élu.
}
\description{
Cette fonction trouve l'élu le plus âgé dans une liste d'élus.
}
\examples{
df <- data.frame(Nom.de.l.élu = c("Dupont", "Martin"),
                 Prénom.de.l.élu = c("Jean", "Sophie"),
                 Date.de.naissance = c("01/01/1950", "15/07/1975"))
}
